Issue #3: Issue with line numbers
When you run the vim-codeclimate commands the line numbers don't
correspond to the codeclimate CLI output (turns out they were being
trimmed to first char).

The issue is with the regex and in particular the hyphen in the regex
which I think is being interpreted as a range. Putting it in a character
set solves the problem.

I haven't been throught the code cliamte code but I tried running the
cli in a number of scenarios and I only ever got output for line numbers
that fit on of the following forms:

* 000(n)
* 000(n)-000(n)

I didn't ever get the equals sign to appear but I've included that in
the character set.

If the CLI output can ever be any other output form then this regex will
need revisiting.

Any questions let me know.

Cheers,

Amo
@sysadmiral